Issue #3: Fixed audio positioning not working.
Audio positioning didn't work when user was in battle and (re)started his TS client.
Apparently it is not possible to attach to a shared memory with write permission if the shared memory already exists.

This is now fixed so that the mod opens the memory with write permission and TS plugin opens it with read permission only.

As a result of this change, version of the plugin cannot be written into the memory anymore (plugin doesn't have write permission). So the version checking functionality is also removed.