Fixed Chocolatey break.
Chocolatey's current version changed and as a result it doesn't add its stuff to $PROFILE anymore. Additionally a new checksum check has been added and some of the packages will refuse to install.

This commit fixes issues that came with those changes. Also, freezed the Chocolatey version to avoid this kind of issues in future.